ecommerce-order-system/order_service/ecommerce-order-system/product_service/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
protobuf==4.21.11
